data-retrieval/requirements.txt
firecrawl-py==2.16.3
python-dotenv==1.0.1
tqdm==4.67.1
requests==2.32.3
geopy==2.4.1
jsonschema==4.23.0
